chore: entropy pass, docs/llms sync, contribution readiness for 0.2.1
- Dedupe parallel-agent redundancy: shared eatAnyPhrase (parse/config),
  canonical stripDateFillers (date/state), dead number-word re-exports
  removed (-17 net LOC, behavior pinned by 912 tests + 6 corpus contracts).
- README/llms.txt/CONTRIBUTING synced to post-wave reality with
  execution-verified examples (treinta y cinco kilos, 三十五公斤,
  las tres menos cuarto, Monday week, 月底); site data refreshed.
- Contribution surface at editor parity: YAML issue forms (bug/feature),
  blank issues off, Discussions enabled with Q&A/Ideas links.
