+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,8 @@
+blank_issues_enabled: false
+contact_links:
+  - name: 💬 Questions & Help
+    url: https://github.com/pascalorg/lingo/discussions/categories/q-a
+    about: Ask questions and get help from the community
+  - name: 💡 Ideas & Discussion
+    url: https://github.com/pascalorg/lingo/discussions/categories/ideas
+    about: Share ideas and discuss features before opening a formal request
