--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,22 +1,0 @@
----
-name: Feature request
-about: A new input form, unit, kind, API, or integration
-labels: enhancement
----
-
-## What should lingo understand or do?
-
-Concrete inputs and expected outputs are the fastest path to a decision:
-
-```text
-"2 stone 4 lb"  →  quantity, mass, 14.5 kg
-```
-
-## Why
-
-The use case — form field, LLM tool argument, MCP server, etc.
-
-## Scope notes (optional)
-
-Anything you already know: prior art, ambiguity hazards ("oz" is mass and
-volume), locale concerns. See `plans/` for how specs are written.
